creating_chatbot/requirements.txt
langchain
langchain-community
flask
python-dotenv
chromadb
openai
tiktoken